backend/requirements.txt
# Force rebuild - 2025-09-25-12-17
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
auth0-python==4.9.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.9.0
boto3==1.35.33
botocore==1.35.99
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.3
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
durationpy==0.10
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
frozenlist==1.6.0
fsspec==2025.5.1
geopy==2.4.1
google-api-core==2.24.2
google-api-python-client==2.146.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
grpcio==1.72.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.8.3
ipython==9.3.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.5.1
json_repair==0.46.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
limits==5.1.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
oauthlib==3.2.2
openai==1.80.0
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-exporter-otlp-proto-http==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-asgi==0.55b0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-util-http==0.55b0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pgvector==0.4.1
platformdirs==4.3.8
posthog==4.2.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyotp==2.9.0
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
pypdf==5.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pillow==11.1.0
pyproject_hooks==1.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.0.7
pdfplumber==0.11.0
python-magic==0.4.27
Pillow==10.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
sendgrid==6.11.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.45.3
stripe==10.12.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.0
zipp==3.22.0
user-agents==2.2.0  # Device detection for session management
